Refactor period map loading logic
Splitting this into functions makes the very long line easier to read.
It also provides a place where the smaller functions can be modified without
affecting the main function. This may be the case if, say, there is a
new resource type which is not in STOR_LONG_DURATION but nevertheless
has LDS-like capabilities which need the Period_Map.