Update EXPANSION_LINES set
Current formulation for partitioning transmission lines to EXPANSION and NO_EXPANSION sets we faulty for stages beyond the first stage, misusing the pMax_Flow_Possible_MW parameter which need not grow between stages for expansion to be permitted. This new, simpler formulation captures the necessary prerequisites for line expansion. This works under testing, both allowing expansion in stages beyond the first now as well as preventing expansion when either of the conditions are not met.