PR to merge Develop with main for version 0.3.3 release (#291)
* Fix multistage DDP bug for output writing

This PR attempts to fix writing of different multi-stage outputs by mentioning the relevant files in GenX.jl and accounting for network expansion

* Update dual_dynamic_programming.jl

Added network expansion condition checking to writing of network expansion output file.

* Time Varying Startup Cost 

This PR attempts to modify the startup cost formulation in commit.jl to reflect the variation with the hours. Instead of the original formulation:

@expression(EP, eCStart[y in COMMIT, t=1:T],(inputs["omega"][t]*inputs["C_Start"][y]*vSTART[y,t]))

it's now written as

@expression(EP, eCStart[y in COMMIT, t=1:T],(inputs["omega"][t]*inputs["C_Start"][y,t]*vSTART[y,t]))

And the doctoring has also been adjusted.

* Added the MGA iterations flag

Added ModelingToGenerateAlternativeIterations flag to GenX_settings.yml and set it to 3 arbitrarily. Tested with Julia 1.6.x, 1.7.x, 1.8.x, and 1.9.x

* One Zone MultiStage MGA Iterations

Flag added to settings

* MGA Iterations flag added to Three Zone MultiStage

In the settings file

* MGA Flag added to ISONE Trizone Multistage

Flag added to the setting file

* MGA iteration flag added to Retrofit Example case

Added to the settings yml file

* Gotten rid of OrdinaryDiffEq from Project.toml (#282)

* Gotten rid of OrdinaryDiffEq from Project.toml

Currently there's no need of this dependency and it ends up creating issues. So, better to get rid of it for now.

* Updated GenX.jl removing OrdinaryDiffEq

Removed the package dependency from GenX.jl

* Added Julia 1.7, 1.8, 1.9, and nightly 

Reinstated the nightly Julia build and added versions 1.7, 1.8, and 1.9 of Julia to enable language upgrades to GenX

* README and documentation update

For version 0.3.3

* Updated how_to_run_genx.md

With the notes of language upgrade

* Updated Project.toml

With the new version patch number

* Addressing issue #280

This PR addresses error flagged in Issue #280. Constraints that limit max charge/discharge rates with or without reserves based on available energy storage capacity did not account for charge or discharge efficiencies. This is corrected here.

* Updated documentation to go with PR #289

updates documentation for storage constraints to match PR #289

* Updates line 205 of storage_all.jl

Inadvertent typing of "*" before "<=" and after "[y,:Eff_Down]"

* Refactor load_network_data for brevity

Add a function to replace the common loading of a ragged column.
General cleanup. Reordering some logic.

* Add option to load the network lines from lists

the lists are in the form of two dataframe columns:
StartNode, EndNode

* Added some documentation

* Add documentation

* Refactor loading scripts (#294)

This does not change the interface at all but makes the code shorter.

Co-authored-by: Chakrabarti, Sambuddha (Sam) <sambuddha.chakrabarti@gmail.com>

Co-authored-by: Jesse Jenkins <jesse.d.jenkins@gmail.com>
Co-authored-by: Jacob Schwartz <jacob@jaschwartz.net>