Add function for extracting matrix from a df
This allows sets of columns to be extracted from a dataframe into a matrix, for
columns which are in any order. Previously the columns needed to be
consecutive and in order.

For example, the ESR_1, ESR_2, ... ESR_N columns in
Energy_share_requirement.csv can now be provided in any order.