Feature: "None" fuel type not required
The "None" fuel type, commonly used for wind, solar, etc, is no longer
required in Fuels_data.csv files. This fixes a point of
confusion for new users (like me...) where it's actually
required to have a column called "None", filled with (presumably) all
zeros. It is automatically added to memory if it doesn't exist.