Remove duplicated license blocks
Rather than keep the license block in each file, we refer readers to
the LICENSE file in the root of the repo. This is done to decrease the
total number of lines of code, to avoid the need to update the copyright
year, and to prevent developers from needing to scroll down.