Refactor main settings configuration
Rather than start with the user settings and add defaults if these user
settings are absent, start with the default and merge the user settings
with these. The outcome is the same, but the code is much cleaner.
This also paves the way to add additional layers of user settings, such
as always using the same solver.

I removed the documentation here (for what setting means what) because it's already present in the
docs/ folder. That should be the canonical source.