Remove MaxCapReq from subsidy revenue
It doesn't make sense to include MaxCapReq as part of a subsidy revenue
in parallel to the MinCapReq constraints. For MinCapReq, this value
represents the subsidies that would be required to induce the system to
reach the specified minimum capacity. Maybe for MaxCapReq there could
be another file instead, that describes the rate at which the total
system cost would decrease were the constraint to be relaxed.