Report constraint's penalty as positive value
For the MinCapReq constraint, a positive value in MinCapReq_prices....
means that the constraint is binding, and that this is the subsidy
required to support the resource so that it reaches the minimum level.

This makes the MaxCapReq constraint output's behavior more symmetric. If the constraint is
minding, the results should show up as a positive number in the
MaxCapReq_prices.[...].csv.