Inclusion of Max Cap Requirement (#365)
Add Maximum Capacity Requirement policy. 

This commit introduces a Policy Module that allows users to limit the total capacity of resources that belong to one or more groups. For example, users can specify that the total capacity of wind and solar resources cannot exceed a certain value. (This forms a multidimensional knapsack problem, for those who are familiar.)

The policy is implemented in a way that parallels the existing Minimum Capacity Requirement policy. It is disabled by default and can be enabled by setting `MaxCapReq: 1` in genx_settings.yml. The policy requires an input file called `Maximum_capacity_requirement.csv` that defines the groups and their maximum capacities. The policy also generates an output file called `MaxCapReq_prices_and_penalties.csv` that reports the penalties associated with the capacity constraints.

---------

Co-authored-by: Jacob Schwartz <jacob@jaschwartz.net>