Abstract REG and RSV terms for therm-commit ramp
A recent change to the THERM_COMMIT ramp expessions added calls to :vREG
and :vRSV variables. However, these variables did not exist at all unless
setup["Reserves"] > 1; and further they would only exist for a subset of
the thermal generators. This would throw an error whenever Reserves was
not 1 or if not all thermal generators with committment had positive Reg_Max and
sv_Max entries.

This change introduces expressions that are defined for all generators y in
THERM_COMMIT, which are 0 if Reserves are off and reference the vREG and
vRSV variables for the correct generators if Reserves are on.

This lets us modify the ramp expression to include these reserve and
regulation terms, while not needing multiple copies of the expression.

This technique could be used to further simplify the Reserve expressions
below.