Reform the interface with TDR by eliminating the OperationWrapping flag (#426)
Eliminate the OperationWrapping flag, and keep the TimeDomainReduction flag.
In the past, certain combinations of the two flags would be required to run full-year cases, or cases with pre-clustered data, or cases where GenX performs the clustering. This was very confusing and led to serious, silent errors.

Now, use TimeDomainReduction=1 if you want GenX to perform clustering and use the results. This will put clustered results in the TDR_Results folder.
Use TimeDomainReduction = 0 if you want an external program (like PowerGenome) to generate reduced time series, or for full-year cases.

Adds functions to validate the time series construction and makes sure that the inputs are self-consistent.

Time series "weights" are always loaded, so expect the `time_weights.csv` output file for all cases.