Update case_runner.jl by removing unnecessary inputs_path
Within the function run_genx_case_simple!(), the AbstractString case is assigned to a newly introduced variable inputs_path. But, this is not necessary any more (maybe its presence is primarily a historic relic from the ancestral GenX.jl). This PR attempts to get rid of that and by doing so, makes the coding style logically connected and in tune with the rest of the coding style.