Fix error from un-initialized sums
A recent JuMP update led to an error where some sums were not
initialized. This would occur in thermal_commit.jl if Up_Time or
Down_Time were 0. This is fixed by setting init=0. Note that this error
did not happen for sums of the form sum(expression[x] for x in S).