Validate resources in generators_data.csv
Checks that the information in generators_data.csv is at least
self-consistent and valid. Right now, this just checks the values in the
columns THERM, LDS, FLEX, HYDRO, VRE, STOR, and LDS.
It makes sure that only one of the mutually-exclusive types is present,
   and that a resource which is LDS is either STOR or HYDRO.
Also, I add the vector of resource-dictionaries (`GenXResource`s) to the
inputs dictionary.

Eventually, relevant data (such as one or more time-series) could be
added to each of these resource dicts.