Backwards compatability
This commit includes a set of changes to ensure that all electrolyzer related inputs are optional and can be excluded without generating runtime errors. This makes the electrolyzer feature backwards compatible.

This commit also changes the documentation notation for the vUSE variable to \Pi to align with storage consumption rather than \Theta which is usually reserved for vP variables.