Rename 'load' to 'demand' in code
'Load' is a confusing term for electrical demand since it's, well, 'overloaded' with the notion of transferring files into memory. 'Demand' is more accessible to non-specialists, and is also a better description of the quantity's role in GenX (an amount of energy desired by users), as demand may not always be fulfilled (if there is non-served energy or flexible demand).
This PR changes 'load' to 'demand', other than keeping the common phrase 'value of lost load' and its abbreviation `Voll`. The changes are fully backward-compatible (and come with `@info` warnings to the user).

Use of `Load` in input files is now deprecated, and support may be dropped in a future version.