Add new tests for seven modules. Include GH workflow. (#563)
This PR adds seven more tests to GenX: three-zone, multi-stage,
electrolyzer, VRE+storage, piecewise_fuel+CO2, TDR, and Method of Morris.
The tests are passed if the value of the objective function is
consistent over time (up to the tolerance of the solver).
The tests for the TDR validate the clustering by computing the rand
index and the mutual information and comparing the output files.
For the Method of Morris module, we are testing if the module is
correctly built and if the output is consistent over time.