Remove functions which worked on Variables
We can't treat Variables and Expressions the same if we want all the functions to be in-place. Using a variable will work but we can't reassign the new expression to the original variable. We could consider other helper functions which take in variables and turn them into expressions. However, in many cases it'll be easier to use add_similar_to_expression!() with the variable as the second argument