Flexible Output Writing (#617)
This commit adds the option of deciding which files to write and which ones to skip. Also, among the time-series files, it gives users the option to choose between writing the hourly outputs versus annual values. 

It does so by adding a new settings parameter called "WriteOutputs" that can be set to annual or full and also introducing a new YAML file called output_settings.yml 

---------

Co-authored-by: Chakrabarti, Sambuddha (Sam) <sambuddha.chakrabarti@gmail.com>