ELCC Capacity Reserve Margin Option
Added the option to give the "CapacityReserveMargin" setting a value of 2, enabling a conventional ELCC-based reserve margin constraint rather than the default dynamic one. In this new formulation the "Capacity_reserve_margnin.csv" input file specifies the absolute instantaneous demand in MW that must be met for each reserve margin constraint (e.g. peak summer or winter demand in a region of interest), and the "Resource_capacity_reserve_margin.csv" input file specifies the fixed ELCC assigned to each resource. There are not separate constraints for each timestep in the ELCC-based formulation.